Initial version
I want to be able to open the Trash with a keyboard shortcut in the
Finder like I can other locations in the Go menu.

This Action allows we to do that, it's the best I could come up with.

The obvious keyboard shortcut to assign this to is command-shift-t,
which is bound to show/hide the tab bar by default, something I do not
use.